refactor(backend/runner): the Cursor adapter's boot owns the proxy interceptors and both roots boot the harnesses before bootstrap (S2 M5, commit 4)
The S1 contract says boot runs before bootstrap resolution and imports the
vendor SDK lazily inside; the M3 adapter did neither, because its static
graph loaded @cursor/sdk (and so connect-node, which snapshots the
node:http2 ESM facade) the moment the module was imported, which is why
the roots still installed the interceptors themselves.

adapter.ts is now the SDK-free lifecycle facade: boot refuses a proxy
endpoint without Config.proxyTokenRef, installs both interceptors over the
ref, emits the interceptors_installed mark where the work happens, runs
assertHttp2ConnectPatched, and only then loads the engine turn with a
dynamic import of the new turn.ts (the former runCursorTurn body, 179 lines
moved by content). Both roots build a bootConfig with the three refs before
bootstrap, call bootHarnesses on it (a new harnesses_booted mark follows;
segment names are data in the cold-start tool), and no longer import any
execute-cursor module. The harness-adapters and registry imports move
ahead of bootstrap and stay dynamic, the runner's dynamic-import-for-order
convention.

harness-boot-order.test.ts proves the contract in a FRESH process, the one
place it can be observed (a vitest worker's facade is frozen before any
test runs): the roots' exact pre-boot imports, then bootHarnesses on a
proxy-mode config resolves; the same after connect-node was imported first
fails with the facade sentence, so the guard inside boot is live.

Comments made true across both interceptors, their tests, the roots, the
adapter table and the hermetic driver; adapter-is-temporal-free names
turn.ts. Hermetic 43/43 twice, goldens byte-identical; kit both halves
green against the fake and the real adapter.
