+++ backend/services/runner/src/runner-manager.ts
@@ -234,9 +234,9 @@ export async function createStigmerRunnerManager(
   // `tokenRef` holds the control-plane token (the host's durable credential,
   // e.g. the desktop's Auth0 token) and is shared with activity clients via
   // config.stigmerTokenRef. The proxy credential (x-stigmer-auth) is managed
-  // separately by a RunnerTokenCoordinator (created after the interceptors are
-  // installed) — see that module for the two-writer rationale and the staleness
-  // history it guards.
+  // separately by a RunnerTokenCoordinator (created after the harnesses boot)
+  // — see that module for the two-writer rationale and the staleness history
+  // it guards.
   const tokenRef = { current: options.stigmerToken ?? null };
   // `runnerTokenRef` mirrors the coordinator's proxy credential for gRPC use:
   // the minted runner token (token_type=embedded_runner) once adopted, tracking
@@ -251,40 +251,23 @@ export async function createStigmerRunnerManager(
 
   assertLlmBackendsPreflight(baseConfig.proxyEndpoint);
 
-  // Install the Cursor SDK interceptors BEFORE resolving Temporal coordinates.
-  // Coordinate discovery dials the control plane via StigmerClient, which loads
-  // @connectrpc/connect-node and snapshots the node:http2 ESM facade on first
-  // import. The HTTP/2 interceptor patches http2.connect and only propagates to
-  // that facade if it runs first, so install MUST precede any connect-node load
-  // (discovery here, and the SDK later). The interceptors depend only on
-  // proxyEndpoint/stigmerToken (already in baseConfig), not resolved coordinates.
-
-  // Install fetch interceptor before any @cursor/sdk imports. Both
-  // interceptors read the proxy credential per request from `runnerTokenRef`
-  // (Config.proxyTokenRef): the control-plane token until a runner token is
-  // minted, the minted token after, exactly the lockstep the coordinator
-  // below maintains in that ref.
-  const { installFetchInterceptor } = await import(
-    "./activities/execute-cursor/fetch-interceptor.js"
-  );
-  installFetchInterceptor({
-    proxyEndpoint: baseConfig.proxyEndpoint ?? undefined,
-    proxyTokenRef: runnerTokenRef,
-  });
-
-  // HTTP/2 interceptor for Connect RPC auth + execution ID injection (BiDi billing)
-  const { installHttp2Interceptor, assertHttp2ConnectPatched } = await import(
-    "./activities/execute-cursor/http2-interceptor.js"
-  );
-  installHttp2Interceptor({
-    proxyEndpoint: baseConfig.proxyEndpoint ?? undefined,
-    proxyTokenRef: runnerTokenRef,
-  });
-  // Fail loudly at boot if a load-order regression left the node:http2 ESM
-  // facade unpatched (otherwise BiDi streams would silently 401). No-op when
-  // the interceptor is unconfigured (no proxy/token).
-  await assertHttp2ConnectPatched();
-  markBoot("interceptors_installed");
+  // Boot the harnesses BEFORE resolving Temporal coordinates. Coordinate
+  // discovery dials the control plane through connect-node, which snapshots
+  // the node:http2 ESM facade on first import; the Cursor adapter's boot
+  // installs the proxy interceptors (the HTTP/2 one patches http2.connect,
+  // which only reaches that facade if it runs first), proves the patch landed,
+  // and only then loads its SDK. The interceptors read the proxy credential
+  // per request from `runnerTokenRef` (Config.proxyTokenRef, bound by
+  // mapManagerOptionsToConfig): the control-plane token until a runner token
+  // is minted, the minted token after, exactly the lockstep the coordinator
+  // below maintains in that ref. Nothing a boot reads depends on the
+  // coordinates, so `baseConfig` carries none yet. The two modules imported
+  // here are connect-free by construction (`harness-adapters.ts`,
+  // `harness/registry.ts`; `__tests__/harness-boot-order.test.ts`).
+  const [{ HARNESS_ADAPTERS }, { adaptersOf, bootHarnesses, releaseHarnessSession, shutdownHarnesses }] =
+    await Promise.all([import("./harness-adapters.js"), import("./harness/registry.js")]);
+  await bootHarnesses(adaptersOf(HARNESS_ADAPTERS), baseConfig);
+  markBoot("harnesses_booted");
 
   // The token coordinator owns the proxy credential (x-stigmer-auth) and its
   // refresh lifecycle. It writes ONE sink, `runnerTokenRef`, which is read per
@@ -304,7 +287,7 @@ export async function createStigmerRunnerManager(
       }),
   });
 
-  // Resolve the runner bootstrap after the http2 patch is in place (discovery
+  // Resolve the runner bootstrap after the harnesses have booted (discovery
   // dials the control plane through connect-node). An explicit address wins;
   // otherwise a token triggers control-plane discovery; otherwise localhost.
   // Must happen before createAllActivities so the activities' Config carries
@@ -541,10 +524,6 @@ export async function createStigmerRunnerManager(
       // (#215). Logged, never thrown: the session's worker is already gone,
       // and failing this IPC command would tell the desktop a removal that
       // did happen did not.
-      const [{ HARNESS_ADAPTERS }, { adaptersOf, releaseHarnessSession }] = await Promise.all([
-        import("./harness-adapters.js"),
-        import("./harness/registry.js"),
-      ]);
       try {
         await releaseHarnessSession(adaptersOf(HARNESS_ADAPTERS), sessionId);
       } catch (err) {
@@ -671,10 +650,6 @@ export async function createStigmerRunnerManager(
       // dispose and whose stdio MCP subprocesses die with them, #215). Logged,
       // never thrown: the drain has already happened, and a teardown failure
       // must not mask a clean stop.
-      const [{ HARNESS_ADAPTERS }, { adaptersOf, shutdownHarnesses }] = await Promise.all([
-        import("./harness-adapters.js"),
-        import("./harness/registry.js"),
-      ]);
       try {
         await shutdownHarnesses(adaptersOf(HARNESS_ADAPTERS));
       } catch (err) {
@@ -758,10 +733,18 @@ export function mapManagerOptionsToConfig(
   };
 }
 
+/**
+ * Dynamically import all activity factories and merge into a single map,
+ * after the harnesses have booted: several of these modules transitively load
+ * `@connectrpc/connect-node` and LangChain, and the Cursor adapter's boot must
+ * have patched `node:http2` before connect-node is first imported (see
+ * createStigmerRunnerManager). The harness rows and the registry were imported
+ * by the factory for the boot; the imports here hit the module cache.
+ */
 async function createAllActivities(config: Config): Promise<WorkerActivities> {
   const [
     { HARNESS_ADAPTERS },
-    { adaptersOf, bootHarnesses, createHarnessActivities },
+    { createHarnessActivities },
     { createDeepAgentActivities },
     { createEnsureThreadActivities },
     { createGenerateSessionSubjectActivities },
@@ -800,11 +783,6 @@ async function createAllActivities(config: Config): Promise<WorkerActivities> {
     import("./activities/promote-task-output.js"),
     import("./activities/attach-session.js"),
   ]);
-
-  // The harness adapters boot here for now (the Cursor adapter's boot keeps
-  // its config slice; the interceptor installs above move into it at S2 M5,
-  // when this call moves ahead of bootstrap resolution, Q-S2-7).
-  await bootHarnesses(adaptersOf(HARNESS_ADAPTERS), config);
 
   return {
     ...(await createHarnessActivities(HARNESS_ADAPTERS, config)),
